Renamed the file containing `JsonModel` to `base`
